first pass at refactoring the code to simplify it
