+++ anomaly/gmm-change-detection/README.md
@@ -6,7 +6,8 @@ This implementation of a GMM-based anomaly detection algorithm was created by Ja
 This algorithm is able to detect changes on environments such as the ISS using 3D point depth cloud data.
 
 # Requirements
-[To-Do]
+pip3 install pulp
+pip3 install scikit-learn
 
 ## Usage
