--- anomaly/gmm-change-detection/scripts/example_code/gmm/gmm_sklearn_example.py
@@ -1,48 +1,0 @@
-# This code was taken from a tutorial from the following website:
-# https://python-course.eu/machine-learning/expectation-maximization-and-gaussian-mixture-models-gmm.php
-
-import matplotlib.pyplot as plt
-from matplotlib import style
-
-style.use("fivethirtyeight")
-import numpy as np
-from scipy.stats import multivariate_normal
-from sklearn.datasets.samples_generator import make_blobs
-from sklearn.mixture import GaussianMixture
-
-# 0. Create dataset
-X, Y = make_blobs(cluster_std=0.5, random_state=20, n_samples=1000, centers=5)
-
-# Stratch dataset to get ellipsoid data
-X = np.dot(X, np.random.RandomState(0).randn(2, 2))
-
-x, y = np.meshgrid(np.sort(X[:, 0]), np.sort(X[:, 1]))
-XY = np.array([x.flatten(), y.flatten()]).T
-
-GMM = GaussianMixture(n_components=5).fit(X)  # Instantiate and fit the model
-print("Converged:", GMM.converged_)  # Check if the model has converged
-means = GMM.means_
-covariances = GMM.covariances_
-
-
-# Predict
-Y = np.array([[0.5], [0.5]])
-prediction = GMM.predict_proba(Y.T)
-print(prediction)
-
-# Plot
-fig = plt.figure(figsize=(10, 10))
-ax0 = fig.add_subplot(111)
-ax0.scatter(X[:, 0], X[:, 1])
-ax0.scatter(Y[0, :], Y[1, :], c="orange", zorder=10, s=100)
-for m, c in zip(means, covariances):
-    multi_normal = multivariate_normal(mean=m, cov=c)
-    ax0.contour(
-        np.sort(X[:, 0]),
-        np.sort(X[:, 1]),
-        multi_normal.pdf(XY).reshape(len(X), len(X)),
-        colors="black",
-        alpha=0.3,
-    )
-    ax0.scatter(m[0], m[1], c="grey", zorder=10, s=100)
-plt.show()
