+++ anomaly/gmm-change-detection/scripts/gmm/artificial_data.py
@@ -4,7 +4,8 @@
 #
 # All rights reserved.
 #
-# The Astrobee platform is licensed under the Apache License, Version 2.0
+# The "ISAAC - Integrated System for Autonomous and Adaptive Caretaking
+# platform" software is licensed under the Apache License, Version 2.0
 # (the "License"); you may not use this file except in compliance with the
 # License. You may obtain a copy of the License at
 #
