+++ anomaly/gmm-change-detection/scripts/gmm/gmm.py
