+++ anomaly/gmm-change-detection/scripts/gmm/ground_truth.py
@@ -1,10 +1,11 @@
-#!/usr/bin/python2
+#!/usr/bin/env python
 # Copyright (c) 2017, United States Government, as represented by the
 # Administrator of the National Aeronautics and Space Administration.
 #
 # All rights reserved.
 #
-# The Astrobee platform is licensed under the Apache License, Version 2.0
+# The "ISAAC - Integrated System for Autonomous and Adaptive Caretaking
+# platform" software is licensed under the Apache License, Version 2.0
 # (the "License"); you may not use this file except in compliance with the
 # License. You may obtain a copy of the License at
 #
