+++ anomaly/gmm-change-detection/setup.py
@@ -1,0 +1,7 @@
+from distutils.core import setup
+
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(packages=["gmm"], package_dir={"": "scripts"})
+
+setup(**d)
